requirements.txt
beautifulsoup4==4.12.3
markdownify==0.11.6
python-slugify==8.0.4
Requests==2.31.0
selenium==4.18.1
